parameterize baseurl and mirrorlist
Also permit passing a parameter repoforge::enabled, an array of strings
representing repositories within repoforge that are to be enabled.  Some
updates to puppet doc to reflect these new parameters.

The basic use cases for this are if you have a local Repoforge mirror and
want to point your clients at it, or if you want to toggle individual
repositories (rpmforge, extras, testing) on and off at will.